Price formatting added;  changed from INT to FLOAT in DB and migrations

+++ database/migrations/2023_11_15_120742_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->timestamps();
             $table->string('artist');
             $table->string('title');
-            $table->integer('price');
+            $table->float('price', 8, 2);
         });
     }
